+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add support for `x**y` with power-of-two `x` or constant `y` (#193)
+
 ### Fixed
 - Fix implementation of assignment expressions (#172, #190)
 - Fix emission of shadow variables for read-write variables in processes